feat: [backend-lint-size-002] - [Bring maintained backend runtime code under the size limits]
