+++ docs/internal/processes/development-guide-relevant-files.md
@@ -85,7 +85,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `.github/workflows/ci.yml` | Repository CI contract | Contributor docs should name this workflow and mirror its root-level validation commands and stated non-deployment scope; the backend coverage job should continue to call `make test-coverage-go` so CI and local runs share the same repo-owned `cmd/gocoveragecheck` policy surface. |
 | `internal/testpath/artifact_contract.go` | Enforced root artifact contract | Root-level factory artifact additions, removals, and redirect stubs are test-enforced here and must stay synchronized with the inventory doc. |
 | `factory/` | Maintainer workflow surface | Contains checked-in operator guidance and active inbox directories that the development guide may reference for workflow-related tasks. |
-| `pkg/api/handlers.go` | Handwritten API decode and validation boundary | `POST /work` and batch decode still need raw-JSON validation at this boundary because generated request structs do not reject unknown keys by themselves, but retired work-request alias and trace-conflict policy should route through `pkg/factory.ValidateCanonicalWorkRequestJSON(...)` here and only be translated into API-owned error wording rather than reimplemented locally. |
+| `pkg/api/handlers.go` plus sibling `pkg/api/handlers_*.go` files | Handwritten API boundary | Keep HTTP handler entrypoints, raw-JSON decode or validation helpers, response projection helpers, and shared API response writers in the same package but split across sibling `handlers_*.go` files when size cleanup is needed; generated request structs still do not reject unknown keys by themselves, so retired work-request alias and trace-conflict policy should continue routing through `pkg/factory.ValidateCanonicalWorkRequestJSON(...)` and only be translated into API-owned error wording rather than reimplemented locally. |
 | `api/codegen_config/client.yaml` | Generated Go client config | The checked-in Go client for backend functional tests is generated from this config and must stay in lockstep with `api/openapi.yaml`, `pkg/generatedclient/`, and `make api-smoke`. |
 | `api/openapi-main.yaml` | Authored OpenAPI source of truth | Public API behavior changes should start here, then refresh the bundled spec and generated Go/UI artifacts together so handwritten route seams and generated parameter binding stay aligned. |
 | `pkg/generatedclient/` | Generated Go API client package | Backend functional tests that need typed client coverage should import this package instead of rebuilding `http.Get` or `http.Post` helpers for OpenAPI-owned routes. |
@@ -107,7 +107,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/factory/event_history_test.go` | Event-history generated payload regression seam | Generated worker and workstation enum-pointer regressions should stay in focused initial-structure package tests here, asserting emitted payload values instead of helper ownership or `tests/functional/**` coverage. |
 | `pkg/interfaces/public_factory_enums.go` | Shared public factory enum normalizers | Shared worker and workstation alias ownership lives here; exported callers must choose explicit strict or permissive helpers instead of duplicating alias tables. |
 | `pkg/interfaces/factory_runtime.go` | Handwritten public work-request boundary structs | Watched-file batch ingestion, generated worker output parsing, and fixture helpers still marshal these structs directly, so their JSON tags must stay aligned with the camelCase OpenAPI contract. |
-| `pkg/cli/dashboard/dashboard.go` | CLI dashboard session-summary boundary | Dashboard session metrics, failed-work detail mapping, and the local fallback from place-category-owned work labels to `Session.DispatchHistory` stay centralized here so CLI/dashboard cleanup work does not duplicate dispatch-history traversal in neighboring packages. |
+| `pkg/cli/dashboard/dashboard.go` plus sibling `pkg/cli/dashboard/dashboard_*.go` files | CLI dashboard session-summary boundary | Keep dashboard render entrypoints, render-data shaping, dispatch-history fallback traversal, and formatting helpers in the same package, but split them into sibling `dashboard_*.go` files when size cleanup is needed so CLI/dashboard work stays behavior-preserving without duplicating traversal or formatting policy in neighboring packages. |
 | `pkg/cli/dashboard/dashboard_test.go` | CLI dashboard behavioral regression surface | Package tests for dashboard rendering and session-summary mapping should assert observable labels and failed-work detail output, especially when `PlaceCategoriesByID` is absent and the CLI must fall back to dispatch history. |
 | `pkg/config/factory_config_mapping.go` | Generated factory-config boundary alias rejection owner | Mirrored legacy alias guards for top-level objects and nested `definition` objects live here; keep one field inventory per boundary type, vary only the path passed into the shared rejection helper, and stop the helper at the documented generated-boundary depth so cleanup work does not broaden alias rejection into deeper nested `definition` chains. |
 | `pkg/config/openapi_factory.go` | Public factory-config normalization and root-guard boundary owner | Path-specific enum normalization and mixed-field rejection for authored `factory.json` root guards belong here so malformed factory guards fail before generated decode with explicit `guards[i].field` messages. |
@@ -360,6 +360,7 @@ This inventory records the checked-in files and directories that the maintainer 
 - Keep customer-visible dashboard failed summaries aligned across `pkg/factory/projections/simple_dashboard_projection.go`, `pkg/cli/dashboardrender/simple_dashboard.go`, and `ui/src/state/factoryTimelineStore.ts`; `failed_count` must count failed work items, matching the population used by `failed_by_work_type`.
 - Active throttle-pause projection must match model-worker lanes against `FactoryWorker.ModelProvider` before executor `Provider`; event-derived topology can carry public uppercase model-provider enums while runtime-defined topology still carries internal executor/provider identifiers.
 - Keep CLI dashboard dispatch-history fallback collection in `pkg/cli/dashboard/dashboard.go` behind one lane-owned helper that selects completed-versus-failed semantics centrally; future cleanup in this area should extend that shared seam instead of reintroducing parallel completed and failed wrapper owners.
+- When `pkg/api` or `pkg/cli/dashboard` hits backend size limits, prefer same-package sibling file splits such as `handlers_*.go` or `dashboard_*.go` that separate entrypoints, decode/validation helpers, projections, and formatting while preserving the exported boundary and tests.
 - In `pkg/factory/runtime/dispatch_result_hook.go`, treat `CompletionDeliveryPlanner` presence as the only synchronous-dispatch mode switch; do not cache a second runtime boolean that can drift from `h.planner`.
 - When local dashboard embed wiring drifts or a legacy `ui/dist_stamp.go` lingers in a working tree, rerun `bun ui/scripts/normalize-dist-output.mjs` or `make ui-build` so the retired stamp file is removed and the generated embed registration is refreshed automatically.
 - In `pkg/config/portable_bundled_files.go`, keep lexical bundled target-path containment in `portableBundledTargetPath(...)` and reserve `validatePortableBundledFilesystemPath(...)` for filesystem-link escape checks along the already-resolved target segments.
