+++ pkg/api/handlers.go
@@ -1,12 +1,8 @@
 package api
 
 import (
-	"bytes"
 	"encoding/base64"
-	"encoding/json"
 	"errors"
-	"fmt"
-	"io"
 	"net/http"
 	"sort"
 	"strings"
@@ -15,10 +11,7 @@ import (
 	"github.com/portpowered/infinite-you/pkg/apisurface"
 	factoryconfig "github.com/portpowered/infinite-you/pkg/config"
 	factorypkg "github.com/portpowered/infinite-you/pkg/factory"
-	"github.com/portpowered/infinite-you/pkg/factory/state"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
-	"github.com/portpowered/infinite-you/pkg/petri"
-	"github.com/portpowered/infinite-you/pkg/workcontent"
 	"github.com/portpowered/infinite-you/pkg/workers"
 	"go.uber.org/zap"
 )
@@ -44,12 +37,6 @@ func (s *Server) SubmitWork(w http.ResponseWriter, r *http.Request) {
 		s.writeError(w, http.StatusBadRequest, "workTypeName is required", "BAD_REQUEST")
 		return
 	}
-	if err := validateGeneratedWorkContentAtPath(req.Content, "content"); err != nil {
-		if message, ok := requestFieldValidationMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-	}
 
 	payload, err := generatedPayloadToRawMessage(req.Payload)
 	if err != nil {
@@ -62,7 +49,7 @@ func (s *Server) SubmitWork(w http.ResponseWriter, r *http.Request) {
 		WorkTypeID:             req.WorkTypeName,
 		CurrentChainingTraceID: stringValue(req.CurrentChainingTraceId),
 		TraceID:                factorypkg.ResolveWorkRequestCurrentChainingTraceID(stringValue(req.CurrentChainingTraceId), stringValue(req.TraceId)),
-		Content:                workcontent.PartsFromGenerated(req.Content),
+		Content:                generatedWorkContentToDomain(req.Content),
 		Payload:                payload,
 		Tags:                   generatedStringMap(req.Tags),
 		Relations:              generatedSubmitRelations(req.Relations),
@@ -555,877 +542,5 @@ func (s *Server) GetEvents(w http.ResponseWriter, r *http.Request) {
 			}
 			flusher.Flush()
 		}
-	}
-}
-
-// --- Helpers ---
-
-func tokenToResponse(t *interfaces.Token, includeHistory bool) factoryapi.TokenResponse {
-	resp := factoryapi.TokenResponse{
-		Id:                       t.ID,
-		PlaceId:                  t.PlaceID,
-		WorkId:                   t.Color.WorkID,
-		WorkType:                 t.Color.WorkTypeID,
-		ChainingTraceDepth:       intPtrIfPositive(t.Color.ChainingTraceDepth),
-		CurrentChainingTraceId:   stringPtrIfNotEmpty(firstNonEmptyString(t.Color.CurrentChainingTraceID, t.Color.TraceID)),
-		PreviousChainingTraceIds: stringSlicePtrCopy(t.Color.PreviousChainingTraceIDs),
-		TraceId:                  t.Color.TraceID,
-		Content:                  domainWorkContentToGeneratedPtr(t.Color.Content),
-		Tags:                     stringMapPtr(t.Color.Tags),
-		CreatedAt:                t.CreatedAt,
-		EnteredAt:                t.EnteredAt,
-	}
-	if t.Color.Name != "" {
-		resp.Name = &t.Color.Name
-	}
-	if len(t.Color.Tags) == 0 {
-		resp.Tags = nil
-	}
-	if includeHistory {
-		resp.History = &factoryapi.TokenHistory{
-			TotalVisits:         integerMapPtr(t.History.TotalVisits),
-			ConsecutiveFailures: integerMapPtr(t.History.ConsecutiveFailures),
-			PlaceVisits:         integerMapPtr(t.History.PlaceVisits),
-			LastError:           stringPtrIfNotEmpty(t.History.LastError),
-		}
-	}
-	return resp
-}
-
-func tokenToWork(t *interfaces.Token, net *state.Net) factoryapi.Work {
-	name := firstNonEmptyString(t.Color.Name, t.Color.WorkID, t.ID)
-	return factoryapi.Work{
-		Name:                     name,
-		WorkId:                   stringPtrIfNotEmpty(t.Color.WorkID),
-		WorkTypeName:             stringPtrIfNotEmpty(t.Color.WorkTypeID),
-		State:                    workStateForToken(t, net),
-		ChainingTraceDepth:       intPtrIfPositive(t.Color.ChainingTraceDepth),
-		CurrentChainingTraceId:   stringPtrIfNotEmpty(firstNonEmptyString(t.Color.CurrentChainingTraceID, t.Color.TraceID)),
-		PreviousChainingTraceIds: stringSlicePtrCopy(t.Color.PreviousChainingTraceIDs),
-		TraceId:                  stringPtrIfNotEmpty(t.Color.TraceID),
-		Content:                  domainWorkContentToGeneratedPtr(t.Color.Content),
-		Tags:                     stringMapPtr(t.Color.Tags),
-	}
-}
-
-func publicWorkNamesByID(tokens map[string]*interfaces.Token) map[string]string {
-	names := make(map[string]string, len(tokens))
-	for _, token := range tokens {
-		if !publicWorkToken(token) || token.Color.WorkID == "" {
-			continue
-		}
-		names[token.Color.WorkID] = firstNonEmptyString(token.Color.Name, token.Color.WorkID, token.ID)
-	}
-	return names
-}
-
-func generatedWorkRelations(token *interfaces.Token, sourceWorkName string, workNamesByID map[string]string) *[]factoryapi.Relation {
-	if token == nil || len(token.Color.Relations) == 0 {
-		return nil
-	}
-
-	relations := make([]factoryapi.Relation, 0, len(token.Color.Relations))
-	for _, relation := range token.Color.Relations {
-		targetWorkName := firstNonEmptyString(workNamesByID[relation.TargetWorkID], relation.TargetWorkID)
-		relations = append(relations, factoryapi.Relation{
-			Type:           factoryapi.RelationType(relation.Type),
-			SourceWorkName: sourceWorkName,
-			TargetWorkName: targetWorkName,
-			TargetWorkId:   stringPtrIfNotEmpty(relation.TargetWorkID),
-			RequiredState:  stringPtrIfNotEmpty(relation.RequiredState),
-		})
-	}
-	return &relations
-}
-
-func workStateForToken(t *interfaces.Token, net *state.Net) *factoryapi.WorkState {
-	if t == nil {
-		return nil
-	}
-	workTypeID, stateName := state.SplitPlaceID(t.PlaceID)
-	if t.Color.WorkTypeID != "" {
-		workTypeID = t.Color.WorkTypeID
-	}
-	if net != nil {
-		if place, ok := net.Places[t.PlaceID]; ok {
-			workTypeID = place.TypeID
-			stateName = place.State
-		}
-	}
-	if stateName == "" {
-		return nil
-	}
-	return &factoryapi.WorkState{
-		Name: stateName,
-		Type: factoryapi.WorkStateType(state.CategoryForState(workTypesFromNet(net), workTypeID, stateName)),
-	}
-}
-
-func workTypesFromNet(net *state.Net) map[string]*state.WorkType {
-	if net == nil {
-		return nil
-	}
-	return net.WorkTypes
-}
-
-func publicWorkToken(token *interfaces.Token) bool {
-	return token != nil &&
-		token.Color.DataType != interfaces.DataTypeResource &&
-		!interfaces.IsSystemTimeToken(token)
-}
-
-func statusFromEngineStateSnapshot(snapshot interfaces.EngineStateSnapshot[petri.MarkingSnapshot, *state.Net]) factoryapi.StatusResponse {
-	categories, resources := categorizeStatusTokens(&snapshot.Marking, snapshot.Topology)
-	return factoryapi.StatusResponse{
-		Categories:    categories,
-		FactoryState:  snapshot.FactoryState,
-		Resources:     resourceUsagePtr(resources),
-		RuntimeStatus: string(snapshot.RuntimeStatus),
-		TotalTokens:   countPublicStatusTokens(&snapshot.Marking),
-	}
-}
-
-func categorizeStatusTokens(marking *petri.MarkingSnapshot, net *state.Net) (factoryapi.StatusCategories, []factoryapi.ResourceUsage) {
-	var categories factoryapi.StatusCategories
-	resourceCounts := make(map[string]int)
-	resourceTotals := resourceTotalsFromTopology(net)
-
-	if marking == nil {
-		return categories, resourceUsage(resourceCounts, resourceTotals)
-	}
-
-	for _, token := range marking.Tokens {
-		if token == nil {
-			continue
-		}
-		if interfaces.IsSystemTimeToken(token) {
-			continue
-		}
-
-		if token.Color.DataType == interfaces.DataTypeResource {
-			resourceID, resourceState := state.SplitPlaceID(token.PlaceID)
-			if _, ok := resourceTotals[resourceID]; !ok {
-				resourceTotals[resourceID]++
-			}
-			if resourceState == interfaces.ResourceStateAvailable {
-				resourceCounts[resourceID]++
-			}
-			continue
-		}
-
-		switch statusStateCategory(net, token.PlaceID) {
-		case state.StateCategoryFailed:
-			categories.Failed++
-		case state.StateCategoryTerminal:
-			categories.Terminal++
-		case state.StateCategoryInitial:
-			categories.Initial++
-		default:
-			categories.Processing++
-		}
-	}
-
-	return categories, resourceUsage(resourceCounts, resourceTotals)
-}
-
-func countPublicStatusTokens(marking *petri.MarkingSnapshot) int {
-	if marking == nil {
-		return 0
-	}
-	count := 0
-	for _, token := range marking.Tokens {
-		if token == nil || interfaces.IsSystemTimeToken(token) {
-			continue
-		}
-		count++
-	}
-	return count
-}
-
-func statusStateCategory(net *state.Net, placeID string) state.StateCategory {
-	if net == nil {
-		return state.StateCategoryProcessing
-	}
-	return net.StateCategoryForPlace(placeID)
-}
-
-func resourceTotalsFromTopology(net *state.Net) map[string]int {
-	totals := make(map[string]int)
-	if net == nil {
-		return totals
-	}
-	for id, resource := range net.Resources {
-		if resource == nil {
-			continue
-		}
-		totals[id] = resource.Capacity
-	}
-	return totals
-}
-
-func resourceUsage(counts map[string]int, totals map[string]int) []factoryapi.ResourceUsage {
-	ids := make([]string, 0, len(totals))
-	for id := range totals {
-		ids = append(ids, id)
-	}
-	sort.Strings(ids)
-
-	resources := make([]factoryapi.ResourceUsage, 0, len(ids))
-	for _, id := range ids {
-		resources = append(resources, factoryapi.ResourceUsage{
-			Available: counts[id],
-			Name:      id,
-			Total:     totals[id],
-		})
-	}
-	return resources
-}
-
-func resourceUsagePtr(values []factoryapi.ResourceUsage) *[]factoryapi.ResourceUsage {
-	if len(values) == 0 {
-		return nil
-	}
-	return &values
-}
-
-func (s *Server) writeJSON(w http.ResponseWriter, status int, v any) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	if err := json.NewEncoder(w).Encode(v); err != nil {
-		s.logger.Error("encode response failed", zap.Error(err))
-	}
-}
-
-func (s *Server) writeError(w http.ResponseWriter, status int, message, code string) {
-	s.writeErrorWithTargets(w, status, message, code, nil)
-}
-
-func (s *Server) writeErrorWithTargets(w http.ResponseWriter, status int, message, code string, targets []factoryapi.ErrorTarget) {
-	var targetPtr *[]factoryapi.ErrorTarget
-	if len(targets) > 0 {
-		targetPtr = &targets
-	}
-	s.writeJSON(w, status, factoryapi.ErrorResponse{
-		Message: message,
-		Family:  errorFamilyForStatus(status),
-		Code:    factoryapi.ErrorResponseCode(code),
-		Targets: targetPtr,
-	})
-}
-
-func errorTarget(kind, id, field string) factoryapi.ErrorTarget {
-	target := factoryapi.ErrorTarget{Kind: kind}
-	if id != "" {
-		target.Id = &id
-	}
-	if field != "" {
-		target.Field = &field
-	}
-	return target
-}
-
-func (s *Server) writeSSEDataJSON(w http.ResponseWriter, v any) error {
-	payload, err := json.Marshal(v)
-	if err != nil {
-		return err
-	}
-	_, err = fmt.Fprintf(w, "data: %s\n\n", payload)
-	return err
-}
-
-func errorFamilyForStatus(status int) factoryapi.ErrorFamily {
-	switch status {
-	case http.StatusBadRequest:
-		return factoryapi.ErrorFamilyBadRequest
-	case http.StatusConflict:
-		return factoryapi.ErrorFamilyConflict
-	case http.StatusNotFound:
-		return factoryapi.ErrorFamilyNotFound
-	default:
-		return factoryapi.ErrorFamilyInternalServerError
-	}
-}
-
-func stringValue(value *string) string {
-	if value == nil {
-		return ""
-	}
-	return *value
-}
-
-func generatedWorkStateName(value *factoryapi.WorkState) string {
-	if value == nil {
-		return ""
-	}
-	return value.Name
-}
-
-func intValue(value *int) int {
-	if value == nil {
-		return 0
-	}
-	return *value
-}
-
-func stringSliceValue(values *[]string) []string {
-	if values == nil {
-		return nil
-	}
-	out := make([]string, len(*values))
-	copy(out, *values)
-	return out
-}
-
-func stringSlicePtrCopy(values []string) *[]string {
-	if len(values) == 0 {
-		return nil
-	}
-	out := append([]string(nil), values...)
-	return &out
-}
-
-func stringPtrIfNotEmpty(value string) *string {
-	if value == "" {
-		return nil
-	}
-	return &value
-}
-
-func integerMapPtr(values map[string]int) *factoryapi.IntegerMap {
-	if len(values) == 0 {
-		return nil
-	}
-	converted := factoryapi.IntegerMap(values)
-	return &converted
-}
-
-func stringMapPtr(values map[string]string) *factoryapi.StringMap {
-	if len(values) == 0 {
-		return nil
-	}
-	converted := factoryapi.StringMap(values)
-	return &converted
-}
-
-func intPtrIfPositive(value int) *int {
-	if value <= 0 {
-		return nil
-	}
-	return &value
-}
-
-func firstNonEmptyString(values ...string) string {
-	for _, value := range values {
-		if value != "" {
-			return value
-		}
-	}
-	return ""
-}
-
-func generatedStringMap(values *factoryapi.StringMap) map[string]string {
-	if values == nil {
-		return nil
-	}
-	return map[string]string(*values)
-}
-
-func generatedSubmitRelations(values *[]factoryapi.SubmitRelation) []interfaces.Relation {
-	if values == nil || len(*values) == 0 {
-		return nil
-	}
-	relations := make([]interfaces.Relation, 0, len(*values))
-	for _, relation := range *values {
-		relations = append(relations, interfaces.Relation{
-			Type:          interfaces.RelationType(relation.Type),
-			TargetWorkID:  relation.TargetWorkId,
-			RequiredState: stringValue(relation.RequiredState),
-		})
-	}
-	return relations
-}
-
-func generatedWorkRequestToDomain(req factoryapi.WorkRequest) (interfaces.WorkRequest, error) {
-	workRequest := interfaces.WorkRequest{
-		RequestID:              req.RequestId,
-		CurrentChainingTraceID: stringValue(req.CurrentChainingTraceId),
-		Type:                   interfaces.WorkRequestType(req.Type),
-	}
-	if req.Works != nil {
-		workRequest.Works = make([]interfaces.Work, 0, len(*req.Works))
-		for i, work := range *req.Works {
-			if err := validateGeneratedWorkContentAtPath(work.Content, fmt.Sprintf("works[%d].content", i)); err != nil {
-				return interfaces.WorkRequest{}, err
-			}
-			workRequest.Works = append(workRequest.Works, interfaces.Work{
-				Name:                     work.Name,
-				WorkID:                   stringValue(work.WorkId),
-				RequestID:                stringValue(work.RequestId),
-				WorkTypeID:               stringValue(work.WorkTypeName),
-				State:                    generatedWorkStateName(work.State),
-				ChainingTraceDepth:       intValue(work.ChainingTraceDepth),
-				CurrentChainingTraceID:   stringValue(work.CurrentChainingTraceId),
-				PreviousChainingTraceIDs: stringSliceValue(work.PreviousChainingTraceIds),
-				TraceID:                  stringValue(work.TraceId),
-				Content:                  workcontent.PartsFromGenerated(work.Content),
-				Payload:                  work.Payload,
-				Tags:                     generatedStringMap(work.Tags),
-			})
-		}
-	}
-	if req.Relations != nil {
-		workRequest.Relations = make([]interfaces.WorkRelation, 0, len(*req.Relations))
-		for _, relation := range *req.Relations {
-			workRequest.Relations = append(workRequest.Relations, interfaces.WorkRelation{
-				Type:           interfaces.WorkRelationType(relation.Type),
-				SourceWorkName: relation.SourceWorkName,
-				TargetWorkName: relation.TargetWorkName,
-				RequiredState:  stringValue(relation.RequiredState),
-			})
-		}
-	}
-	return workRequest, nil
-}
-
-func domainWorkContentToGeneratedPtr(parts []interfaces.WorkContentPart) *factoryapi.WorkContent {
-	if len(parts) == 0 {
-		return nil
-	}
-	content := make(factoryapi.WorkContent, 0, len(parts))
-	for _, part := range parts {
-		var generated factoryapi.WorkContentPart
-		switch part.Type {
-		case interfaces.WorkContentPartTypeText:
-			if err := generated.FromWorkTextContentPart(factoryapi.WorkTextContentPart{
-				Type: factoryapi.WorkContentPartTypeText,
-				Text: part.Text,
-			}); err != nil {
-				continue
-			}
-		case interfaces.WorkContentPartTypeImage:
-			if err := generated.FromWorkImageContentPart(factoryapi.WorkImageContentPart{
-				Type: factoryapi.WorkContentPartTypeImage,
-				File: part.File,
-			}); err != nil {
-				continue
-			}
-		default:
-			continue
-		}
-		content = append(content, generated)
-	}
-	if len(content) == 0 {
-		return nil
-	}
-	return &content
-}
-
-func validateGeneratedWorkContentAtPath(content *factoryapi.WorkContent, fieldPath string) error {
-	if content == nil || len(*content) == 0 {
-		return nil
-	}
-
-	for i, part := range *content {
-		pathPrefix := fmt.Sprintf("%s[%d].", fieldPath, i)
-		textPart, textErr := part.AsWorkTextContentPart()
-		if textErr == nil && textPart.Type == factoryapi.WorkContentPartTypeText {
-			continue
-		}
-
-		imagePart, imageErr := part.AsWorkImageContentPart()
-		if imageErr == nil && imagePart.Type == factoryapi.WorkContentPartTypeImage {
-			continue
-		}
-
-		return requestFieldValidationError{message: fmt.Sprintf("%stype must be one of text or image", pathPrefix)}
-	}
-	return nil
-}
-
-type requestFieldValidationError struct {
-	message string
-}
-
-func (e requestFieldValidationError) Error() string {
-	return e.message
-}
-
-func requestFieldValidationMessage(err error) (string, bool) {
-	var validationErr requestFieldValidationError
-	if errors.As(err, &validationErr) {
-		return validationErr.message, true
-	}
-	return "", false
-}
-
-func decodeSubmitWorkRequestBody(body io.Reader) (factoryapi.SubmitWorkJSONRequestBody, error) {
-	data, err := io.ReadAll(body)
-	if err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-
-	var req factoryapi.SubmitWorkJSONRequestBody
-	if err := json.Unmarshal(data, &req); err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-	if err := validateCanonicalWorkRequestJSONForAPI(data); err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-
-	var fields map[string]json.RawMessage
-	if err := json.Unmarshal(data, &fields); err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-	if err := validateWorkContentField(fields, ""); err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-	if strings.TrimSpace(req.Name) == "" {
-		return factoryapi.SubmitWorkJSONRequestBody{}, requestFieldValidationError{message: "name is required"}
-	}
-	return req, nil
-}
-
-func decodeWorkRequestBody(body io.Reader) (factoryapi.UpsertWorkRequestJSONRequestBody, error) {
-	data, err := io.ReadAll(body)
-	if err != nil {
-		return factoryapi.UpsertWorkRequestJSONRequestBody{}, err
-	}
-
-	var req factoryapi.UpsertWorkRequestJSONRequestBody
-	if err := json.Unmarshal(data, &req); err != nil {
-		return factoryapi.UpsertWorkRequestJSONRequestBody{}, err
-	}
-	if err := validateCanonicalWorkRequestJSONForAPI(data); err != nil {
-		return factoryapi.UpsertWorkRequestJSONRequestBody{}, err
-	}
-
-	if req.Works == nil || len(*req.Works) == 0 {
-		return req, nil
-	}
-
-	var rawRequest struct {
-		Works []map[string]json.RawMessage `json:"works"`
-	}
-	if err := json.Unmarshal(data, &rawRequest); err != nil {
-		return factoryapi.UpsertWorkRequestJSONRequestBody{}, err
-	}
-
-	for i := range *req.Works {
-		if i >= len(rawRequest.Works) {
-			return req, nil
-		}
-		if err := validateWorkContentField(rawRequest.Works[i], fmt.Sprintf("works[%d].", i)); err != nil {
-			return factoryapi.UpsertWorkRequestJSONRequestBody{}, err
-		}
-	}
-	return req, nil
-}
-
-func decodeNamedFactoryBody(body io.Reader) (factoryapi.CreateFactoryJSONRequestBody, error) {
-	data, err := io.ReadAll(body)
-	if err != nil {
-		return factoryapi.CreateFactoryJSONRequestBody{}, err
-	}
-
-	decoder := json.NewDecoder(bytes.NewReader(data))
-	decoder.DisallowUnknownFields()
-
-	var req factoryapi.CreateFactoryJSONRequestBody
-	if err := decoder.Decode(&req); err != nil {
-		return factoryapi.CreateFactoryJSONRequestBody{}, err
-	}
-	if err := decoder.Decode(&struct{}{}); err != io.EOF {
-		if err == nil {
-			return factoryapi.CreateFactoryJSONRequestBody{}, requestFieldValidationError{message: "request payload must contain one JSON object"}
-		}
-		return factoryapi.CreateFactoryJSONRequestBody{}, err
-	}
-	return req, nil
-}
-
-func decodeSaveEditableFactoryDefinitionBody(body io.Reader) (factoryapi.SaveEditableCurrentFactoryDefinitionJSONRequestBody, error) {
-	data, err := io.ReadAll(body)
-	if err != nil {
-		return factoryapi.SaveEditableCurrentFactoryDefinitionJSONRequestBody{}, err
-	}
-
-	decoder := json.NewDecoder(bytes.NewReader(data))
-	decoder.DisallowUnknownFields()
-
-	var req factoryapi.SaveEditableCurrentFactoryDefinitionJSONRequestBody
-	if err := decoder.Decode(&req); err != nil {
-		return factoryapi.SaveEditableCurrentFactoryDefinitionJSONRequestBody{}, err
-	}
-	if err := decoder.Decode(&struct{}{}); err != io.EOF {
-		if err == nil {
-			return factoryapi.SaveEditableCurrentFactoryDefinitionJSONRequestBody{}, requestFieldValidationError{message: "request payload must contain one JSON object"}
-		}
-		return factoryapi.SaveEditableCurrentFactoryDefinitionJSONRequestBody{}, err
-	}
-	return req, nil
-}
-func decodePromptTemplateValidationRequestBody(body io.Reader) (factoryapi.ValidateCurrentFactoryWorkstationPromptTemplateJSONRequestBody, error) {
-	data, err := io.ReadAll(body)
-	if err != nil {
-		return factoryapi.ValidateCurrentFactoryWorkstationPromptTemplateJSONRequestBody{}, err
-	}
-
-	dec := json.NewDecoder(bytes.NewReader(data))
-	dec.DisallowUnknownFields()
-
-	var req factoryapi.ValidateCurrentFactoryWorkstationPromptTemplateJSONRequestBody
-	if err := dec.Decode(&req); err != nil {
-		return factoryapi.ValidateCurrentFactoryWorkstationPromptTemplateJSONRequestBody{}, err
-	}
-	if err := ensureSingleJSONObject(dec); err != nil {
-		return factoryapi.ValidateCurrentFactoryWorkstationPromptTemplateJSONRequestBody{}, err
-	}
-
-	return req, nil
-}
-
-func ensureSingleJSONObject(dec *json.Decoder) error {
-	if err := dec.Decode(&struct{}{}); err != io.EOF {
-		if err == nil {
-			return requestFieldValidationError{message: "request payload must contain one JSON object"}
-		}
-		return err
-	}
-	return nil
-}
-
-func validateCanonicalWorkRequestJSONForAPI(data []byte) error {
-	if err := factorypkg.ValidateCanonicalWorkRequestJSON(data); err != nil {
-		return translateCanonicalWorkRequestValidationError(err)
-	}
-	return nil
-}
-
-func currentFactoryWorkstation(factory factoryapi.Factory, workstationName string) (factoryapi.Workstation, bool) {
-	if factory.Workstations == nil {
-		return factoryapi.Workstation{}, false
-	}
-	for _, workstation := range *factory.Workstations {
-		if workstation.Name == workstationName || stringValue(workstation.Id) == workstationName {
-			return workstation, true
-		}
-	}
-	return factoryapi.Workstation{}, false
-}
-
-func promptTemplateContractResponse(contract workers.PromptTemplateContract) factoryapi.PromptTemplateContract {
-	availableVariables := make([]factoryapi.PromptTemplateVariableReference, 0, len(contract.AvailableVariables))
-	for _, reference := range contract.AvailableVariables {
-		availableVariables = append(availableVariables, factoryapi.PromptTemplateVariableReference{
-			Category:    factoryapi.PromptTemplateVariableReferenceCategory(reference.Category),
-			Description: reference.Description,
-			Example:     reference.Example,
-			Path:        reference.Path,
-		})
-	}
-	unavailablePatterns := make([]factoryapi.PromptTemplateUnavailableAccessPattern, 0, len(contract.UnavailableAccessPatterns))
-	for _, pattern := range contract.UnavailableAccessPatterns {
-		unavailablePatterns = append(unavailablePatterns, factoryapi.PromptTemplateUnavailableAccessPattern{
-			Example: pattern.Example,
-			Path:    pattern.Path,
-			Reason:  pattern.Reason,
-		})
-	}
-	return factoryapi.PromptTemplateContract{
-		AvailableVariables:        availableVariables,
-		InputCount:                contract.InputCount,
-		UnavailableAccessPatterns: unavailablePatterns,
-	}
-}
-
-func promptTemplateValidationResultResponse(result workers.PromptTemplateValidationResult) factoryapi.PromptTemplateValidationResult {
-	diagnostics := make([]factoryapi.PromptTemplateDiagnostic, 0, len(result.Diagnostics))
-	for _, diagnostic := range result.Diagnostics {
-		diagnostics = append(diagnostics, factoryapi.PromptTemplateDiagnostic{
-			EndOffset:   diagnostic.EndOffset,
-			Kind:        factoryapi.PromptTemplateDiagnosticKind(diagnostic.Kind),
-			Message:     diagnostic.Message,
-			Path:        diagnostic.Path,
-			SourceText:  diagnostic.SourceText,
-			StartOffset: diagnostic.StartOffset,
-		})
-	}
-	return factoryapi.PromptTemplateValidationResult{
-		Diagnostics: diagnostics,
-		Valid:       result.Valid,
-	}
-}
-func validateWorkContentField(fields map[string]json.RawMessage, prefix string) error {
-	contentRaw, ok := fields["content"]
-	if !ok {
-		return nil
-	}
-
-	var partPayloads []json.RawMessage
-	if err := json.Unmarshal(contentRaw, &partPayloads); err != nil {
-		return requestFieldValidationError{message: fmt.Sprintf("%scontent must be an array", prefix)}
-	}
-	for i, payload := range partPayloads {
-		var partFields map[string]json.RawMessage
-		if err := json.Unmarshal(payload, &partFields); err != nil {
-			return requestFieldValidationError{message: fmt.Sprintf("%scontent[%d] must be an object", prefix, i)}
-		}
-		if _, err := validatedRawWorkContentPart(partFields, fmt.Sprintf("%scontent[%d].", prefix, i)); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func validatedRawWorkContentPart(fields map[string]json.RawMessage, prefix string) (interfaces.WorkContentPart, error) {
-	typeRaw, ok := fields["type"]
-	if !ok {
-		return interfaces.WorkContentPart{}, requestFieldValidationError{message: fmt.Sprintf("%stype is required", prefix)}
-	}
-
-	var partType string
-	if err := json.Unmarshal(typeRaw, &partType); err != nil || partType == "" {
-		return interfaces.WorkContentPart{}, requestFieldValidationError{message: fmt.Sprintf("%stype must be a non-empty string", prefix)}
-	}
-
-	switch interfaces.WorkContentPartType(partType) {
-	case interfaces.WorkContentPartTypeText:
-		if err := requireOnlyFields(fields, prefix, "type", "text"); err != nil {
-			return interfaces.WorkContentPart{}, err
-		}
-		textRaw, ok := fields["text"]
-		if !ok {
-			return interfaces.WorkContentPart{}, requestFieldValidationError{message: fmt.Sprintf("%stext is required for text content parts", prefix)}
-		}
-		var text string
-		if err := json.Unmarshal(textRaw, &text); err != nil {
-			return interfaces.WorkContentPart{}, requestFieldValidationError{message: fmt.Sprintf("%stext must be a string", prefix)}
-		}
-		return interfaces.WorkContentPart{Type: interfaces.WorkContentPartTypeText, Text: text}, nil
-	case interfaces.WorkContentPartTypeImage:
-		if err := requireOnlyFields(fields, prefix, "type", "file"); err != nil {
-			return interfaces.WorkContentPart{}, err
-		}
-		fileRaw, ok := fields["file"]
-		if !ok {
-			return interfaces.WorkContentPart{}, requestFieldValidationError{message: fmt.Sprintf("%sfile is required for image content parts", prefix)}
-		}
-		var file string
-		if err := json.Unmarshal(fileRaw, &file); err != nil || file == "" {
-			return interfaces.WorkContentPart{}, requestFieldValidationError{message: fmt.Sprintf("%sfile must be a non-empty string", prefix)}
-		}
-		return interfaces.WorkContentPart{Type: interfaces.WorkContentPartTypeImage, File: file}, nil
-	default:
-		return interfaces.WorkContentPart{}, requestFieldValidationError{message: fmt.Sprintf("%stype must be one of text or image", prefix)}
-	}
-}
-
-func requireOnlyFields(fields map[string]json.RawMessage, prefix string, allowed ...string) error {
-	allowedSet := make(map[string]struct{}, len(allowed))
-	for _, field := range allowed {
-		allowedSet[field] = struct{}{}
-	}
-	for field := range fields {
-		if _, ok := allowedSet[field]; ok {
-			continue
-		}
-		return requestFieldValidationError{message: fmt.Sprintf("%s%s is not supported", prefix, field)}
-	}
-	return nil
-}
-
-func applyStableTraceToWorkRequest(req *interfaces.WorkRequest) {
-	if req == nil || len(req.Works) == 0 {
-		return
-	}
-	traceID := ""
-	if req.CurrentChainingTraceID != "" {
-		traceID = req.CurrentChainingTraceID
-	}
-	if traceID == "" {
-		for _, work := range req.Works {
-			if work.CurrentChainingTraceID != "" {
-				traceID = work.CurrentChainingTraceID
-				break
-			}
-			if work.TraceID != "" {
-				traceID = work.TraceID
-				break
-			}
-		}
-	}
-	if traceID == "" {
-		traceID = "trace-" + req.RequestID
-	}
-	if req.CurrentChainingTraceID == "" {
-		req.CurrentChainingTraceID = traceID
-	}
-	for i := range req.Works {
-		if req.Works[i].CurrentChainingTraceID == "" {
-			if req.Works[i].TraceID != "" {
-				req.Works[i].CurrentChainingTraceID = req.Works[i].TraceID
-			} else {
-				req.Works[i].CurrentChainingTraceID = traceID
-			}
-		}
-		if req.Works[i].TraceID == "" {
-			req.Works[i].TraceID = req.Works[i].CurrentChainingTraceID
-		}
-	}
-}
-
-func generatedPayloadToRawMessage(payload any) (json.RawMessage, error) {
-	if payload == nil {
-		return nil, nil
-	}
-	return json.Marshal(payload)
-}
-
-func submitWorkBadRequestMessage(err error) (string, bool) {
-	if err == nil {
-		return "", false
-	}
-	message := err.Error()
-	if strings.HasPrefix(message, "work_request:") {
-		return submitWorkTypeNameMessage(message), true
-	}
-	if strings.Contains(message, "unknown work type") || strings.Contains(message, "work type") && strings.Contains(message, "not found") {
-		return submitWorkTypeNameMessage(message), true
-	}
-	return "", false
-}
-
-func submitWorkTypeNameMessage(message string) string {
-	message = strings.ReplaceAll(message, "work_type_name", "workTypeName")
-	message = strings.ReplaceAll(message, "work_type_id", "workTypeName")
-	if strings.Contains(message, "work type name") {
-		return message
-	}
-	return strings.ReplaceAll(message, "work type", "work type name")
-}
-
-func translateCanonicalWorkRequestValidationError(err error) error {
-	if err == nil {
-		return nil
-	}
-
-	message := err.Error()
-	message = strings.TrimPrefix(message, "work request batch ")
-	message = strings.ReplaceAll(message, " uses retired work_type_id field; use workTypeName", ".work_type_id is not supported; use workTypeName")
-	message = strings.ReplaceAll(message, " uses retired target_state field; use state", ".target_state is not supported; use state")
-	if strings.HasPrefix(message, "works[") && strings.Contains(message, "] ") {
-		message = strings.Replace(message, "] ", "].", 1)
-	}
-	if strings.HasSuffix(message, ".work_type_id is not supported; use workTypeName") ||
-		strings.HasSuffix(message, ".target_state is not supported; use state") {
-		return requestFieldValidationError{message: message}
-	}
-	switch message {
-	case "uses retired work_type_id field; use workTypeName":
-		return requestFieldValidationError{message: "work_type_id is not supported; use workTypeName"}
-	case "uses retired target_state field; use state":
-		return requestFieldValidationError{message: "target_state is not supported; use state"}
-	default:
-		return requestFieldValidationError{message: message}
 	}
 }
