+++ pkg/api/handlers_response.go
@@ -1,0 +1,145 @@
+package api
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+
+	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+	"go.uber.org/zap"
+)
+
+func (s *Server) writeJSON(w http.ResponseWriter, status int, v any) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(status)
+	if err := json.NewEncoder(w).Encode(v); err != nil {
+		s.logger.Error("encode response failed", zap.Error(err))
+	}
+}
+
+func (s *Server) writeError(w http.ResponseWriter, status int, message, code string) {
+	s.writeErrorWithTargets(w, status, message, code, nil)
+}
+
+func (s *Server) writeErrorWithTargets(w http.ResponseWriter, status int, message, code string, targets []factoryapi.ErrorTarget) {
+	var targetPtr *[]factoryapi.ErrorTarget
+	if len(targets) > 0 {
+		targetPtr = &targets
+	}
+	s.writeJSON(w, status, factoryapi.ErrorResponse{
+		Message: message,
+		Family:  errorFamilyForStatus(status),
+		Code:    factoryapi.ErrorResponseCode(code),
+		Targets: targetPtr,
+	})
+}
+
+func errorTarget(kind, id, field string) factoryapi.ErrorTarget {
+	target := factoryapi.ErrorTarget{Kind: kind}
+	if id != "" {
+		target.Id = &id
+	}
+	if field != "" {
+		target.Field = &field
+	}
+	return target
+}
+
+func (s *Server) writeSSEDataJSON(w http.ResponseWriter, v any) error {
+	payload, err := json.Marshal(v)
+	if err != nil {
+		return err
+	}
+	_, err = fmt.Fprintf(w, "data: %s\n\n", payload)
+	return err
+}
+
+func errorFamilyForStatus(status int) factoryapi.ErrorFamily {
+	switch status {
+	case http.StatusBadRequest:
+		return factoryapi.ErrorFamilyBadRequest
+	case http.StatusConflict:
+		return factoryapi.ErrorFamilyConflict
+	case http.StatusNotFound:
+		return factoryapi.ErrorFamilyNotFound
+	default:
+		return factoryapi.ErrorFamilyInternalServerError
+	}
+}
+
+func stringValue(value *string) string {
+	if value == nil {
+		return ""
+	}
+	return *value
+}
+
+func generatedWorkStateName(value *factoryapi.WorkState) string {
+	if value == nil {
+		return ""
+	}
+	return value.Name
+}
+
+func intValue(value *int) int {
+	if value == nil {
+		return 0
+	}
+	return *value
+}
+
+func stringSliceValue(values *[]string) []string {
+	if values == nil {
+		return nil
+	}
+	out := make([]string, len(*values))
+	copy(out, *values)
+	return out
+}
+
+func stringSlicePtrCopy(values []string) *[]string {
+	if len(values) == 0 {
+		return nil
+	}
+	out := append([]string(nil), values...)
+	return &out
+}
+
+func stringPtrIfNotEmpty(value string) *string {
+	if value == "" {
+		return nil
+	}
+	return &value
+}
+
+func integerMapPtr(values map[string]int) *factoryapi.IntegerMap {
+	if len(values) == 0 {
+		return nil
+	}
+	converted := factoryapi.IntegerMap(values)
+	return &converted
+}
+
+func stringMapPtr(values map[string]string) *factoryapi.StringMap {
+	if len(values) == 0 {
+		return nil
+	}
+	converted := factoryapi.StringMap(values)
+	return &converted
+}
+
+func intPtrIfPositive(value int) *int {
+	if value <= 0 {
+		return nil
+	}
+	return &value
+}
+
+func firstNonEmptyString(values ...string) string {
+	for _, value := range values {
+		if value != "" {
+			return value
+		}
+	}
+	return ""
+}
